+++ internal/controller/datadogagent/component/clusteragent/default_test.go
@@ -249,5 +249,21 @@ func clusterAgentDefaultEnvVars(dda *datadoghqv2alpha1.DatadogAgent) []corev1.En
 			Name:  "DD_INSTRUMENTATION_INSTALL_ID",
 			Value: "",
 		},
+		{
+			Name:  "DD_CLUSTER_AGENT_SERVICE_ACCOUNT_NAME",
+			Value: "foo-cluster-agent",
+		},
+		{
+			Name:  "DD_AGENT_DAEMONSET",
+			Value: "foo-agent",
+		},
+		{
+			Name:  "DD_CLUSTER_AGENT_DEPLOYMENT",
+			Value: "foo-cluster-agent",
+		},
+		{
+			Name:  "DD_DATADOG_AGENT_CUSTOM_RESOURCE",
+			Value: "foo",
+		},
 	}
 }
