+++ internal/controller/datadogagent/component/clusteragent/envvar.go
@@ -6,7 +6,11 @@
 package clusteragent
 
 const (
-	DDAuthTokenFilePath      = "DD_AUTH_TOKEN_FILE_PATH"
-	DDKubeResourcesNamespace = "DD_KUBE_RESOURCES_NAMESPACE"
-	DDPodName                = "DD_POD_NAME"
+	DDAuthTokenFilePath              = "DD_AUTH_TOKEN_FILE_PATH"
+	DDKubeResourcesNamespace         = "DD_KUBE_RESOURCES_NAMESPACE"
+	DDPodName                        = "DD_POD_NAME"
+	DDClusterAgentServiceAccountName = "DD_CLUSTER_AGENT_SERVICE_ACCOUNT_NAME"
+	DDAgentDaemonSet                 = "DD_AGENT_DAEMONSET"
+	DDClusterAgentDeployment         = "DD_CLUSTER_AGENT_DEPLOYMENT"
+	DDDatadogAgentCustomResource     = "DD_DATADOG_AGENT_CUSTOM_RESOURCE"
 )
