+++ internal/controller/datadogagent/controller_reconcile_dca.go
@@ -22,6 +22,7 @@ import (
 	datadoghqv2alpha1 "github.com/DataDog/datadog-operator/api/datadoghq/v2alpha1"
 	apiutils "github.com/DataDog/datadog-operator/api/utils"
 	"github.com/DataDog/datadog-operator/internal/controller/datadogagent/common"
+	"github.com/DataDog/datadog-operator/internal/controller/datadogagent/component"
 	componentdca "github.com/DataDog/datadog-operator/internal/controller/datadogagent/component/clusteragent"
 	"github.com/DataDog/datadog-operator/internal/controller/datadogagent/feature"
 	"github.com/DataDog/datadog-operator/internal/controller/datadogagent/global"
@@ -147,7 +148,7 @@ func (r *Reconciler) cleanupOldDCADeployments(ctx context.Context, logger logr.L
 		kubernetes.AppKubernetesManageByLabelKey:   "datadog-operator",
 		kubernetes.AppKubernetesPartOfLabelKey:     object.NewPartOfLabelValue(dda).String(),
 	}
-	deploymentName := getDeploymentNameFromDCA(dda)
+	deploymentName := component.GetDeploymentNameFromDatadogAgent(dda)
 	deploymentList := appsv1.DeploymentList{}
 	if err := r.client.List(ctx, &deploymentList, matchLabels); err != nil {
 		return err
@@ -161,16 +162,4 @@ func (r *Reconciler) cleanupOldDCADeployments(ctx context.Context, logger logr.L
 	}
 
 	return nil
-}
-
-// getDeploymentNameFromDCA returns the expected DCA deployment name based on
-// the DDA name and clusterAgent name override
-func getDeploymentNameFromDCA(dda *datadoghqv2alpha1.DatadogAgent) string {
-	deploymentName := componentdca.GetClusterAgentName(dda)
-	if componentOverride, ok := dda.Spec.Override[datadoghqv2alpha1.ClusterAgentComponentName]; ok {
-		if componentOverride.Name != nil && *componentOverride.Name != "" {
-			deploymentName = *componentOverride.Name
-		}
-	}
-	return deploymentName
 }
