+++ internal/controller/datadogagent/feature/enabledefault/envvar.go
@@ -11,7 +11,4 @@ const (
 	// InstallInfoToolVersion is used by the Operator to override the tool
 	// version value in the Agent's install info
 	InstallInfoToolVersion = "DD_TOOL_VERSION"
-
-	DDKubernetesResourcesLabelsAsTags      = "DD_KUBERNETES_RESOURCES_LABELS_AS_TAGS"
-	DDKubernetesResourcesAnnotationsAsTags = "DD_KUBERNETES_RESOURCES_ANNOTATIONS_AS_TAGS"
 )
