+++ pkg/constants/utils_test.go
@@ -154,122 +154,3 @@ func TestServiceAccountNameOverride(t *testing.T) {
 		})
 	}
 }
-
-func TestServiceAccountAnnotationOverride(t *testing.T) {
-	customServiceAccount := "fake"
-	customServiceAccountAnnotations := map[string]string{
-		"eks.amazonaws.com/role-arn": "arn:aws:iam::123456789012:role/datadog-role",
-		"really.important":           "annotation",
-	}
-	ddaName := "test-dda"
-	tests := []struct {
-		name string
-		dda  *v2alpha1.DatadogAgent
-		want map[v2alpha1.ComponentName]map[string]interface{}
-	}{
-		{
-			name: "custom serviceaccount annotations for dda, dca and clc",
-			dda: &v2alpha1.DatadogAgent{
-				ObjectMeta: v1.ObjectMeta{
-					Name: ddaName,
-				},
-				Spec: v2alpha1.DatadogAgentSpec{
-					Override: map[v2alpha1.ComponentName]*v2alpha1.DatadogAgentComponentOverride{
-						v2alpha1.ClusterAgentComponentName: {
-							ServiceAccountName:        &customServiceAccount,
-							ServiceAccountAnnotations: customServiceAccountAnnotations,
-						},
-						v2alpha1.ClusterChecksRunnerComponentName: {
-							ServiceAccountAnnotations: customServiceAccountAnnotations,
-						},
-						v2alpha1.NodeAgentComponentName: {
-							ServiceAccountAnnotations: customServiceAccountAnnotations,
-						},
-					},
-				},
-			},
-			want: map[v2alpha1.ComponentName]map[string]interface{}{
-				v2alpha1.ClusterAgentComponentName: {
-					"name":        customServiceAccount,
-					"annotations": customServiceAccountAnnotations,
-				},
-				v2alpha1.NodeAgentComponentName: {
-					"name":        fmt.Sprintf("%s-%s", ddaName, DefaultAgentResourceSuffix),
-					"annotations": customServiceAccountAnnotations,
-				},
-				v2alpha1.ClusterChecksRunnerComponentName: {
-					"name":        fmt.Sprintf("%s-%s", ddaName, DefaultClusterChecksRunnerResourceSuffix),
-					"annotations": customServiceAccountAnnotations,
-				},
-			},
-		},
-		{
-			name: "custom serviceaccount annotations for dca",
-			dda: &v2alpha1.DatadogAgent{
-				ObjectMeta: v1.ObjectMeta{
-					Name: ddaName,
-				},
-				Spec: v2alpha1.DatadogAgentSpec{
-					Override: map[v2alpha1.ComponentName]*v2alpha1.DatadogAgentComponentOverride{
-						v2alpha1.ClusterAgentComponentName: {
-							ServiceAccountName:        &customServiceAccount,
-							ServiceAccountAnnotations: customServiceAccountAnnotations,
-						},
-					},
-				},
-			},
-			want: map[v2alpha1.ComponentName]map[string]interface{}{
-				v2alpha1.NodeAgentComponentName: {
-					"name":        fmt.Sprintf("%s-%s", ddaName, DefaultAgentResourceSuffix),
-					"annotations": map[string]string{},
-				},
-				v2alpha1.ClusterAgentComponentName: {
-					"name":        customServiceAccount,
-					"annotations": customServiceAccountAnnotations,
-				},
-				v2alpha1.ClusterChecksRunnerComponentName: {
-					"name":        fmt.Sprintf("%s-%s", ddaName, DefaultClusterChecksRunnerResourceSuffix),
-					"annotations": map[string]string{},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			res := map[v2alpha1.ComponentName]map[string]interface{}{
-				v2alpha1.NodeAgentComponentName: {
-					"name":        GetAgentServiceAccount(tt.dda),
-					"annotations": GetAgentServiceAccountAnnotations(tt.dda),
-				},
-				v2alpha1.ClusterChecksRunnerComponentName: {
-					"name":        GetClusterChecksRunnerServiceAccount(tt.dda),
-					"annotations": GetClusterChecksRunnerServiceAccountAnnotations(tt.dda),
-				},
-				v2alpha1.ClusterAgentComponentName: {
-					"name":        GetClusterAgentServiceAccount(tt.dda),
-					"annotations": GetClusterAgentServiceAccountAnnotations(tt.dda),
-				},
-			}
-			for componentName, sa := range tt.want {
-				if res[componentName]["name"] != sa["name"] {
-					t.Errorf("Service Account Override Name error = %v, want %v", res[componentName], tt.want[componentName])
-				}
-				if !mapsEqual(res[componentName]["annotations"].(map[string]string), sa["annotations"].(map[string]string)) {
-					t.Errorf("Service Account Override Annotation error = %v, want %v", res[componentName], tt.want[componentName])
-				}
-			}
-		})
-	}
-}
-
-func mapsEqual(a, b map[string]string) bool {
-	if len(a) != len(b) {
-		return false
-	}
-	for key, value := range a {
-		if bValue, ok := b[key]; !ok || value != bValue {
-			return false
-		}
-	}
-	return true
-}
